back-end/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
pydantic
mplfinance
pandas
